Move fixtures to trait
