+++ tests/Common/FixtureTrait.php
@@ -2,12 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Data\Tests\Common\Reader\ReaderWithFilter;
+namespace Yiisoft\Data\Tests\Common;
 
-use Yiisoft\Data\Reader\DataReaderInterface;
-use Yiisoft\Data\Tests\TestCase;
-
-abstract class BaseReaderWithFilterTestCase extends TestCase
+trait FixtureTrait
 {
     protected static $fixtures = [
         ['number' => 1, 'email' => 'foo@bar', 'balance' => 10.25, 'born_at' => null],
@@ -16,8 +13,6 @@ abstract class BaseReaderWithFilterTestCase extends TestCase
         ['number' => 4, 'email' => 'the@best', 'balance' => 500.0, 'born_at' => null],
         ['number' => 5, 'email' => 'test@test', 'balance' => 42.0, 'born_at' => '1990-01-01'],
     ];
-
-    abstract protected function getReader(): DataReaderInterface;
 
     protected function assertFixtures(array $expectedFixtureIndexes, array $actualFixtures): void
     {
