+++ tests/Common/Reader/ReaderWithFilter/BaseReaderWithBetweenTestCase.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace Yiisoft\Data\Tests\Common\Reader\ReaderWithFilter;
 
 use Yiisoft\Data\Reader\Filter\Between;
+use Yiisoft\Data\Tests\Common\Reader\BaseReaderTestCase;
 
-abstract class BaseReaderWithBetweenTestCase extends BaseReaderWithFilterTestCase
+abstract class BaseReaderWithBetweenTestCase extends BaseReaderTestCase
 {
     public function testWithReader(): void
     {
