+++ tests/Common/Reader/ReaderWithFilter/BaseReaderWithEqualsNullTestCase.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace Yiisoft\Data\Tests\Common\Reader\ReaderWithFilter;
 
 use Yiisoft\Data\Reader\Filter\EqualsNull;
+use Yiisoft\Data\Tests\Common\Reader\BaseReaderTestCase;
 
-abstract class BaseReaderWithEqualsNullTestCase extends BaseReaderWithFilterTestCase
+abstract class BaseReaderWithEqualsNullTestCase extends BaseReaderTestCase
 {
     public function testWithReader(): void
     {
