+++ tests/Common/Reader/ReaderWithFilter/BaseReaderWithEqualsTestCase.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace Yiisoft\Data\Tests\Common\Reader\ReaderWithFilter;
 
 use Yiisoft\Data\Reader\Filter\Equals;
+use Yiisoft\Data\Tests\Common\Reader\BaseReaderTestCase;
 
-abstract class BaseReaderWithEqualsTestCase extends BaseReaderWithFilterTestCase
+abstract class BaseReaderWithEqualsTestCase extends BaseReaderTestCase
 {
     public function testWithReader(): void
     {
