+++ tests/Common/Reader/ReaderWithFilter/BaseReaderWithGreaterThanOrEqualTestCase.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace Yiisoft\Data\Tests\Common\Reader\ReaderWithFilter;
 
 use Yiisoft\Data\Reader\Filter\GreaterThanOrEqual;
+use Yiisoft\Data\Tests\Common\Reader\BaseReaderTestCase;
 
-abstract class BaseReaderWithGreaterThanOrEqualTestCase extends BaseReaderWithFilterTestCase
+abstract class BaseReaderWithGreaterThanOrEqualTestCase extends BaseReaderTestCase
 {
     public function testWithReader(): void
     {
