+++ tests/Common/Reader/ReaderWithFilter/BaseReaderWithInTestCase.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace Yiisoft\Data\Tests\Common\Reader\ReaderWithFilter;
 
 use Yiisoft\Data\Reader\Filter\In;
+use Yiisoft\Data\Tests\Common\Reader\BaseReaderTestCase;
 
-abstract class BaseReaderWithInTestCase extends BaseReaderWithFilterTestCase
+abstract class BaseReaderWithInTestCase extends BaseReaderTestCase
 {
     public function testWithReader(): void
     {
