+++ tests/Common/Reader/ReaderWithFilter/BaseReaderWithLessThanOrEqualTestCase.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace Yiisoft\Data\Tests\Common\Reader\ReaderWithFilter;
 
 use Yiisoft\Data\Reader\Filter\LessThanOrEqual;
+use Yiisoft\Data\Tests\Common\Reader\BaseReaderTestCase;
 
-abstract class BaseReaderWithLessThanOrEqualTestCase extends BaseReaderWithFilterTestCase
+abstract class BaseReaderWithLessThanOrEqualTestCase extends BaseReaderTestCase
 {
     public function testWithReader(): void
     {
