+++ tests/Common/Reader/ReaderWithFilter/BaseReaderWithLessThanTestCase.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace Yiisoft\Data\Tests\Common\Reader\ReaderWithFilter;
 
 use Yiisoft\Data\Reader\Filter\LessThan;
+use Yiisoft\Data\Tests\Common\Reader\BaseReaderTestCase;
 
-abstract class BaseReaderWithLessThanTestCase extends BaseReaderWithFilterTestCase
+abstract class BaseReaderWithLessThanTestCase extends BaseReaderTestCase
 {
     public function testWithReader(): void
     {
